--- data/duke.txt
@@ -1,1 +1,0 @@
-E # 0 # asjdhasjhd #  16:00:15 #  16:00:15
