+++ src/main/java/Commands/MarkDonenessCommand.java
@@ -1,9 +1,9 @@
 package Commands;
 
 import Exceptions.DukeException;
-import TaskPackages.TaskList;
-import Utility.Storage;
-import Utility.Ui;
+import Tasks.TaskList;
+import Utilities.Storage;
+import Utilities.Ui;
 
 public class MarkDonenessCommand extends Command {
