+++ src/main/java/Commands/ToDoCommand.java
@@ -1,8 +1,8 @@
 package Commands;
 
-import TaskPackages.TaskList;
-import Utility.Storage;
-import Utility.Ui;
+import Tasks.TaskList;
+import Utilities.Storage;
+import Utilities.Ui;
 
 public class ToDoCommand extends TaskCommand {
