+++ src/main/java/Tasks/Deadline.java
@@ -1,4 +1,4 @@
-package TaskPackages;
+package Tasks;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
