+++ src/main/java/Tasks/TaskList.java
@@ -1,4 +1,4 @@
-package TaskPackages;
+package Tasks;
 
 import java.util.ArrayList;
