+++ src/main/java/Tasks/Todo.java
@@ -1,4 +1,4 @@
-package TaskPackages;
+package Tasks;
 
 public class Todo extends Task {
