+++ src/main/java/Utilities/DukeEnum.java
@@ -1,4 +1,4 @@
-package Utility;
+package Utilities;
 
 public enum DukeEnum {
