+++ src/main/java/Utilities/Parser.java
@@ -1,4 +1,4 @@
-package Utility;
+package Utilities;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
