+++ src/main/java/Utilities/Ui.java
@@ -1,4 +1,4 @@
-package Utility;
+package Utilities;
 
 import java.util.Scanner;
